refactor: separate code to make it cleaner
