+++ project-e-dukate-frontend/src/components/FilterButton/index.ts
@@ -1,0 +1,1 @@
+export * from './FilterButton';
