+++ project-e-dukate-frontend/src/components/Payments/EditableCurrencyField.tsx
@@ -1,0 +1,36 @@
+import React from "react";
+import { Box, Typography } from "@mui/material";
+import { TextField } from "../TextField";
+
+interface EditableCurrencyFieldProps {
+  value: string;
+  onChange: (value: string) => void;
+}
+
+export const EditableCurrencyField: React.FC<EditableCurrencyFieldProps> = ({
+  value,
+  onChange,
+}) => {
+  return (
+    <Box
+      sx={{
+        display: "flex",
+        alignItems: "center",
+        justifyContent: "center",
+        gap: 1,
+      }}
+    >
+      <TextField
+        value={value}
+        onChange={onChange}
+        sx={{
+          "& .MuiInputBase-root": { height: "auto" },
+          width: "70px",
+        }}
+        type="number"
+        placeholder="0"
+      />
+      <Typography sx={{ color: "black" }}>bs.</Typography>
+    </Box>
+  );
+};
