+++ project-e-dukate-frontend/src/components/Payments/index.ts
@@ -1,0 +1,4 @@
+export * from './PaymentTable';
+export * from './PaymentTableRow';
+export * from './PaymentFilters';
+export * from './EditableCurrencyField';
